feat: integrate framework image agent for image generation
