+++ blocklets/core/api/src/providers/models.ts
@@ -1,339 +1,16 @@
-import type { Agent } from 'node:https';
-
-import { AnthropicChatModel } from '@aigne/anthropic';
-import { BedrockChatModel } from '@aigne/bedrock';
-import {
-  AIGNE,
-  AgentResponseStream,
-  ChatModel,
-  ChatModelOptions,
-  ChatModelOutput,
-  Message,
-  isAgentResponseDelta,
-} from '@aigne/core';
-import { DeepSeekChatModel } from '@aigne/deepseek';
-import { DoubaoChatModel } from '@aigne/doubao';
-import { GeminiChatModel } from '@aigne/gemini';
-import { OllamaChatModel } from '@aigne/ollama';
-import { OpenRouterChatModel } from '@aigne/open-router';
-import { OpenAIChatModel } from '@aigne/openai';
-import type { OpenAIChatModelOptions } from '@aigne/openai';
-import { PoeChatModel } from '@aigne/poe';
-import { XAIChatModel } from '@aigne/xai';
-import { getModelNameWithProvider } from '@api/libs/ai-provider';
+import { availableModels as availableChatModels, availableImageModels } from '@aigne/aigne-hub';
+import { AIGNE, ChatModelOptions, ChatModelOutput, ImageModelOptions } from '@aigne/core';
+import type { OpenAIChatModelOptions, OpenAIImageModelOptions } from '@aigne/openai';
 import logger from '@api/libs/logger';
 import AiCredential from '@api/store/models/ai-credential';
 import AiProvider from '@api/store/models/ai-provider';
-import { ConfigError, ConfigErrorType } from '@blocklet/aigne-hub/api';
-import { ChatCompletionChunk, ChatCompletionInput, ChatCompletionResponse } from '@blocklet/aigne-hub/api/types';
-import { getRemoteBaseUrl } from '@blocklet/aigne-hub/api/utils/util';
+import { ChatCompletionInput, ChatCompletionResponse } from '@blocklet/aigne-hub/api/types';
 import { CustomError } from '@blocklet/error';
-import { NodeHttpHandler, streamCollector } from '@smithy/node-http-handler';
-import { HttpsProxyAgent } from 'https-proxy-agent';
 import { pick } from 'lodash';
-import { joinURL } from 'ufo';
 
-import { Config } from '../libs/env';
-
-function convertToFrameworkMessages(
-  messages: ChatCompletionInput['messages']
-): import('@aigne/core').ChatModelInputMessage[] {
-  return messages.map((message): import('@aigne/core').ChatModelInputMessage => {
-    switch (message.role) {
-      case 'system':
-        return {
-          role: 'system' as const,
-          content: message.content,
-        };
-
-      case 'user':
-        return {
-          role: 'user' as const,
-          content:
-            typeof message.content === 'string'
-              ? message.content
-              : message.content.map((item) => {
-                  if (item.type === 'text') {
-                    return { type: 'text', text: item.text };
-                  }
-                  if (item.type === 'image_url') {
-                    return { type: 'image_url', url: item.imageUrl.url };
-                  }
-                  return item;
-                }),
-        };
-
-      case 'assistant':
-        return {
-          role: 'agent' as const,
-          content: message.content,
-          toolCalls: message.toolCalls?.map((call) => ({
-            id: call.id,
-            type: 'function' as const,
-            function: {
-              name: call.function.name,
-              arguments: call.function.arguments as unknown as Message,
-            },
-          })),
-        };
-
-      case 'tool':
-        return {
-          role: 'tool' as const,
-          content: message.content,
-          toolCallId: message.toolCallId,
-        };
-
-      default:
-        // @ts-ignore
-        throw new CustomError(400, `Unknown message role: ${message.role}`);
-    }
-  });
-}
-
-const providers = {
-  openai: 'openai',
-  anthropic: 'anthropic',
-  bedrock: 'bedrock',
-  deepseek: 'deepseek',
-  google: 'google',
-  ollama: 'ollama',
-  openrouter: 'openrouter',
-  xai: 'xai',
-  doubao: 'doubao',
-  poe: 'poe',
-} as const;
-
-type AIProvider = keyof typeof providers;
-
-function availableModels(): {
-  name: string;
-  provider: AIProvider;
-  create: (options: {
-    model?: string;
-    modelOptions?: ChatModelOptions;
-    clientOptions?: OpenAIChatModelOptions['clientOptions'];
-  }) => ChatModel;
-}[] {
-  const { httpsProxy } = Config;
-  const proxy = ['HTTPS_PROXY', 'https_proxy', 'HTTP_PROXY', 'http_proxy', 'ALL_PROXY', 'all_proxy']
-    .map((i) => process.env[i] ?? (httpsProxy === undefined ? undefined : httpsProxy === i ? httpsProxy : undefined))
-    .filter(Boolean)[0];
-
-  const httpAgent = proxy ? (new HttpsProxyAgent(proxy) as Agent) : undefined;
-  const clientOptions = { fetchOptions: { agent: httpAgent } };
-
-  return [
-    {
-      name: OpenAIChatModel.name,
-      provider: providers.openai,
-      create: (params) => new OpenAIChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: AnthropicChatModel.name,
-      provider: providers.anthropic,
-      create: (params) => new AnthropicChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: BedrockChatModel.name,
-      provider: providers.bedrock,
-      create: (params) =>
-        new BedrockChatModel({
-          ...params,
-          clientOptions: {
-            requestHandler: NodeHttpHandler.create({ httpAgent, httpsAgent: httpAgent }),
-            streamCollector,
-          },
-        }),
-    },
-    {
-      name: DeepSeekChatModel.name,
-      provider: providers.deepseek,
-      create: (params) => new DeepSeekChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: GeminiChatModel.name,
-      provider: providers.google,
-      create: (params) => new GeminiChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: OllamaChatModel.name,
-      provider: providers.ollama,
-      create: (params) => new OllamaChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: OpenRouterChatModel.name,
-      provider: providers.openrouter,
-      create: (params) => new OpenRouterChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: XAIChatModel.name,
-      provider: providers.xai,
-      create: (params) => new XAIChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: DoubaoChatModel.name,
-      provider: providers.doubao,
-      create: (params) => new DoubaoChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: PoeChatModel.name,
-      provider: providers.poe,
-      create: (params) => new PoeChatModel({ ...params, clientOptions }),
-    },
-  ];
-}
-
-const currentApiKeyIndex: { [key in AIProvider]?: number } = {};
-const apiKeys: { [key in AIProvider]: () => string[] } = {
-  google: () => Config.geminiApiKey,
-  openai: () => Config.openaiApiKey,
-  openrouter: () => Config.openRouterApiKey,
-  anthropic: () => Config.anthropicApiKey,
-  deepseek: () => Config.deepseekApiKey,
-  bedrock: () => Config.awsAccessKeyId,
-  ollama: () => Config.ollamaApiKey,
-  xai: () => Config.xaiApiKey,
-  doubao: () => Config.doubaoApiKey,
-  poe: () => Config.poeApiKey,
-};
-
-const aigneHubConfigProviderUrl = async () => {
-  const url = await getRemoteBaseUrl(process.env?.BLOCKLET_AIGNE_API_URL || '').catch(
-    () => process.env?.BLOCKLET_AIGNE_API_URL
-  );
-  const errorMessage = `Please config provider in ${joinURL(url || '', 'config/ai-config/providers')}`;
-  return errorMessage;
-};
-
-async function getAIApiKey(company: AIProvider) {
-  currentApiKeyIndex[company] ??= 0;
-
-  const index = currentApiKeyIndex[company]!++;
-  const keys = apiKeys[company]?.();
-
-  const key = keys?.[index % keys.length];
-
-  if (!key) {
-    const configUrl = await aigneHubConfigProviderUrl();
-    throw new ConfigError(ConfigErrorType.MISSING_API_KEY, configUrl);
-  }
-
-  return { apiKey: key };
-}
-
-async function getBedrockConfig() {
-  currentApiKeyIndex.bedrock ??= 0;
-
-  const index = currentApiKeyIndex.bedrock!++;
-  const accessKeyIds = Config.awsAccessKeyId;
-  const secretAccessKeys = Config.awsSecretAccessKey;
-  const regions = Config.awsRegion;
-
-  const accessKeyId = accessKeyIds?.[index % accessKeyIds.length];
-  const secretAccessKey = secretAccessKeys?.[index % secretAccessKeys.length];
-  const region = regions?.[index % regions.length];
-
-  if (!accessKeyId || !secretAccessKey || !region) {
-    const configUrl = await aigneHubConfigProviderUrl();
-    throw new ConfigError(ConfigErrorType.MISSING_API_KEY, configUrl);
-  }
-
-  return { accessKeyId, secretAccessKey, region };
-}
-
-const BASE_URL_CONFIG_MAP = {
-  openai: () => Config.openaiBaseURL,
-  anthropic: () => Config.anthropicBaseURL,
-  ollama: () => Config.ollamaBaseURL,
-} as const;
-
-export async function loadModel(
-  model: string,
-  {
-    provider,
-    modelOptions,
-    clientOptions,
-    req,
-  }: {
-    provider?: string;
-    modelOptions?: ChatModelOptions;
-    clientOptions?: OpenAIChatModelOptions['clientOptions'];
-    req?: any; // Express Request with modelCallContext
-  } = {}
-) {
-  const models = availableModels();
-  const m = models.find((m) => provider && m.provider.toLowerCase().includes(provider.toLowerCase()));
-
-  if (!m)
-    throw new CustomError(
-      404,
-      `Provider ${provider} model ${model} not found, Please check the model name and provider.`
-    );
-
-  const params: {
-    apiKey?: string;
-    baseURL?: string;
-    accessKeyId?: string;
-    secretAccessKey?: string;
-    region?: string;
-    modelOptions?: ChatModelOptions;
-    clientOptions?: OpenAIChatModelOptions['clientOptions'];
-  } = await getProviderCredentials(m.provider, {
-    modelCallContext: req?.modelCallContext,
-    model,
-  });
-
-  if (modelOptions) {
-    params.modelOptions = modelOptions;
-  }
-
-  if (clientOptions) {
-    params.clientOptions = clientOptions;
-  }
-
-  return m.create({ ...params, model });
-}
-
-export async function getModelAndProviderId(model: string) {
-  let { providerName, modelName } = getModelNameWithProvider(model);
-
-  const getDefaultProvider = () => {
-    if (model.toLowerCase().startsWith('gemini')) return 'google';
-    if (model.toLowerCase().startsWith('gpt')) return 'openai';
-    if (model.toLowerCase().startsWith('openrouter')) return 'openrouter';
-
-    if (!providerName || !modelName) {
-      throw new CustomError(
-        400,
-        'The model format is incorrect. Please use {provider}/{model}, for example: openai/gpt-4o or anthropic/claude-3-5-sonnet'
-      );
-    }
-
-    return '';
-  };
-
-  if (!providerName) {
-    providerName = getDefaultProvider();
-    modelName = model;
-  }
-
-  const provider = await AiProvider.findOne({ where: { name: providerName } });
-  return { providerId: provider?.id, modelName, providerName };
-}
-
-export const getModel = async (
-  input: ChatCompletionInput & Required<Pick<ChatCompletionInput, 'model'>>,
-  options?: {
-    modelOptions?: ChatModelOptions;
-    clientOptions?: OpenAIChatModelOptions['clientOptions'];
-    req?: any;
-  }
-) => {
-  const { modelName: model, providerName: provider } = await getModelAndProviderId(input.model);
-  const m = await loadModel(model, { provider, ...options });
-  return m;
-};
+import { AIProvider } from './constants';
+import { BASE_URL_CONFIG_MAP, aigneHubConfigProviderUrl, getAIApiKey, getBedrockConfig } from './keys';
+import { adaptStreamToOldFormat, convertToFrameworkMessages, getModelAndProviderId } from './util';
 
 export async function getProviderCredentials(
   provider: string,
@@ -419,6 +96,7 @@ export async function getProviderCredentials(
     region: providerRecord?.region,
   };
 }
+
 export async function chatCompletionByFrameworkModel(
   input: ChatCompletionInput & Required<Pick<ChatCompletionInput, 'model'>>,
   userDid?: string,
@@ -445,65 +123,130 @@ export async function chatCompletionByFrameworkModel(
   return adaptStreamToOldFormat(response);
 }
 
-export async function* adaptStreamToOldFormat(
-  stream: AgentResponseStream<ChatModelOutput>
-): AsyncGenerator<ChatCompletionResponse> {
-  const toolCalls: ChatCompletionChunk['delta']['toolCalls'] = [];
-  const role: ChatCompletionChunk['delta']['role'] = 'assistant';
+async function loadModel(
+  model: string,
+  {
+    provider,
+    modelOptions,
+    clientOptions,
+    req,
+  }: {
+    provider?: string;
+    modelOptions?: ChatModelOptions;
+    clientOptions?: OpenAIChatModelOptions['clientOptions'];
+    req?: any; // Express Request with modelCallContext
+  } = {}
+) {
+  const models = availableChatModels();
+  const providerName = provider?.toLowerCase().replace(/-/g, '') || '';
 
-  for await (const chunk of stream) {
-    if (isAgentResponseDelta(chunk)) {
-      const { delta } = chunk;
-
-      if (delta.json?.toolCalls && Array.isArray(delta.json.toolCalls)) {
-        for (const call of delta.json.toolCalls) {
-          toolCalls.push(call);
-        }
-      }
-
-      if (delta.json?.json) {
-        yield {
-          delta: {
-            role,
-            content: JSON.stringify(delta.json.json),
-          },
-        };
-      }
-
-      if (delta.text?.text || delta.json?.toolCalls) {
-        yield {
-          delta: {
-            role,
-            content: delta.text?.text,
-            toolCalls:
-              Array.isArray(toolCalls) && toolCalls.length > 0
-                ? toolCalls.map((call) => ({
-                    ...call,
-                    function: {
-                      name: call.function?.name,
-                      arguments:
-                        call.function?.arguments && typeof call.function.arguments === 'object'
-                          ? JSON.stringify(call.function.arguments)
-                          : call.function?.arguments,
-                    },
-                  }))
-                : [],
-          },
-        };
-      }
-
-      if (delta.json?.usage) {
-        const { inputTokens = 0, outputTokens = 0 } =
-          (delta.json.usage as { inputTokens: number; outputTokens: number }) || {};
-
-        yield {
-          usage: {
-            promptTokens: inputTokens,
-            completionTokens: outputTokens,
-            totalTokens: inputTokens + outputTokens,
-          },
-        };
-      }
+  const m = models.find((m) => {
+    if (typeof m.name === 'string') {
+      return m.name.toLowerCase().includes(providerName);
     }
+    return m.name.some((n) => n.toLowerCase().includes(providerName));
+  });
+
+  if (!m)
+    throw new CustomError(
+      404,
+      `Provider ${provider} model ${model} not found, Please check the model name and provider.`
+    );
+
+  const params: {
+    apiKey?: string;
+    baseURL?: string;
+    accessKeyId?: string;
+    secretAccessKey?: string;
+    region?: string;
+    modelOptions?: ChatModelOptions;
+    clientOptions?: OpenAIChatModelOptions['clientOptions'];
+  } = await getProviderCredentials(providerName, {
+    modelCallContext: req?.modelCallContext,
+    model,
+  });
+
+  if (modelOptions) {
+    params.modelOptions = modelOptions;
   }
+
+  if (clientOptions) {
+    params.clientOptions = clientOptions;
+  }
+
+  return m.create({ ...params, model });
 }
+
+export const getModel = async (
+  input: ChatCompletionInput & Required<Pick<ChatCompletionInput, 'model'>>,
+  options?: {
+    modelOptions?: ChatModelOptions;
+    clientOptions?: OpenAIChatModelOptions['clientOptions'];
+    req?: any;
+  }
+) => {
+  const { modelName: model, providerName: provider } = await getModelAndProviderId(input.model);
+  const m = await loadModel(model, { provider, ...options });
+  return m;
+};
+
+const loadImageModel = async (
+  model: string,
+  {
+    provider,
+    modelOptions,
+    clientOptions,
+    req,
+  }: {
+    provider?: string;
+    modelOptions?: ImageModelOptions;
+    clientOptions?: OpenAIImageModelOptions['clientOptions'];
+    req?: any; // Express Request with modelCallContext
+  } = {}
+) => {
+  const models = availableImageModels();
+  const providerName = (provider || '').toLowerCase() === 'google' ? 'gemini' : provider?.toLowerCase();
+  const m = models.find((m) => providerName && m.name.toLowerCase().includes(providerName.toLowerCase()));
+
+  if (!m)
+    throw new CustomError(
+      404,
+      `Provider ${provider} model ${model} not found, Please check the model name and provider.`
+    );
+
+  const params: {
+    apiKey?: string;
+    baseURL?: string;
+    accessKeyId?: string;
+    secretAccessKey?: string;
+    region?: string;
+    modelOptions?: ImageModelOptions;
+    clientOptions?: OpenAIImageModelOptions['clientOptions'];
+  } = await getProviderCredentials(provider!, {
+    modelCallContext: req?.modelCallContext,
+    model,
+  });
+
+  if (modelOptions) {
+    params.modelOptions = modelOptions;
+  }
+
+  if (clientOptions) {
+    params.clientOptions = clientOptions;
+  }
+
+  return m.create({ ...params, model });
+};
+
+export const getImageModel = async (
+  input: { model: string },
+  options?: {
+    modelOptions?: ImageModelOptions;
+    clientOptions?: OpenAIImageModelOptions['clientOptions'];
+    req?: any;
+  }
+) => {
+  const { modelName: model, providerName: provider } = await getModelAndProviderId(input.model);
+  const m = await loadImageModel(model, { provider, ...options });
+  return m;
+};
