+++ blocklets/core/package.json
@@ -32,19 +32,11 @@
   "dependencies": {
     "@abtnode/cron": "^1.16.48",
     "@abtnode/util": "^1.16.48",
-    "@aigne/anthropic": "^0.11.4",
-    "@aigne/bedrock": "^0.9.4",
-    "@aigne/core": "^1.51.0",
-    "@aigne/deepseek": "^0.7.22",
-    "@aigne/doubao": "^1.0.16",
-    "@aigne/gemini": "^0.9.4",
-    "@aigne/observability-api": "^0.9.0",
-    "@aigne/ollama": "^0.7.22",
-    "@aigne/open-router": "^0.7.22",
-    "@aigne/openai": "^0.11.4",
-    "@aigne/poe": "^1.0.2",
-    "@aigne/transport": "^0.14.2",
-    "@aigne/xai": "^0.7.22",
+    "@aigne/aigne-hub": "^0.7.0",
+    "@aigne/core": "^1.55.1",
+    "@aigne/observability-api": "^0.10.1",
+    "@aigne/openai": "^0.13.0",
+    "@aigne/transport": "^0.14.9",
     "@arcblock/did": "^1.22.2",
     "@arcblock/did-connect-js": "^1.22.2",
     "@arcblock/did-connect-react": "^3.1.22",
